+++ app/src/main/java/com/sopt/now/network/ApiFactory.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.data.signup
+package com.sopt.now.network
 
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import com.sopt.now.BuildConfig
